fix SAFE_MODE restrictions on temp folder write access
modified:   gvexport/gvexport.php

Without this fix, the system temp folder was used as a
temp folder, but SAFE_MODE restriction prevented write
access, making GVExport useless.
Now a write check was added, and a last attempt was
made by using the index/ folder as a temp folder (as
for it the write access must be granted for PGV to
work).